Rm word native
