+++ products/jbrowse-cli/src/commands/create.ts
@@ -1,14 +1,13 @@
 import fs from 'fs'
-
-import { Args, Flags } from '@oclif/core'
+import { parseArgs } from 'util'
 import decompress from 'decompress'
 
-import JBrowseCommand from '../base'
+import NativeCommand from '../native-base'
 import fetch from '../fetchWithProxy'
 
 const fsPromises = fs.promises
 
-export default class Create extends JBrowseCommand {
+export default class CreateNative extends NativeCommand {
   static description = 'Downloads and installs the latest JBrowse 2 release'
 
   static examples = [
@@ -28,110 +27,134 @@ export default class Create extends JBrowseCommand {
     '$ jbrowse create --listVersions',
   ]
 
-  static args = {
-    localPath: Args.string({
-      required: true,
-      description: 'Location where JBrowse 2 will be installed',
-    }),
-  }
-
-  static flags = {
-    help: Flags.help({ char: 'h' }),
-    force: Flags.boolean({
-      char: 'f',
-      description:
-        'Overwrites existing JBrowse 2 installation if present in path',
-    }),
-    // will need to account for pagination once there is a lot of releases
-    listVersions: Flags.boolean({
-      char: 'l',
-      description: 'Lists out all versions of JBrowse 2',
-    }),
-    branch: Flags.string({
-      description: 'Download a development build from a named git branch',
-    }),
-    nightly: Flags.boolean({
-      description: 'Download the latest development build from the main branch',
-    }),
-    url: Flags.string({
-      char: 'u',
-      description: 'A direct URL to a JBrowse 2 release',
-    }),
-    tag: Flags.string({
-      char: 't',
-      description:
-        'Version of JBrowse 2 to install. Format is v1.0.0.\nDefaults to latest',
-    }),
-  }
-
   async run() {
-    const { args: runArgs, flags: runFlags } = await this.parse(Create)
-    const { localPath: argsPath } = runArgs as { localPath: string }
-    this.debug(`Want to install path at: ${argsPath}`)
+    const { values: flags, positionals } = parseArgs({
+      args: process.argv.slice(3), // Skip node, script, and command name
+      options: {
+        help: {
+          type: 'boolean',
+          short: 'h',
+          default: false,
+        },
+        force: {
+          type: 'boolean',
+          short: 'f',
+          default: false,
+        },
+        listVersions: {
+          type: 'boolean',
+          short: 'l',
+          default: false,
+        },
+        branch: {
+          type: 'string',
+        },
+        nightly: {
+          type: 'boolean',
+          default: false,
+        },
+        url: {
+          type: 'string',
+          short: 'u',
+        },
+        tag: {
+          type: 'string',
+          short: 't',
+        },
+      },
+      allowPositionals: true,
+    })
 
-    const { force, url, listVersions, tag, branch, nightly } = runFlags
+    if (flags.help) {
+      this.showHelp()
+      return
+    }
 
-    if (listVersions) {
+    if (flags.listVersions) {
       const versions = (await this.fetchGithubVersions()).map(
         version => version.tag_name,
       )
-      this.log(`All JBrowse versions:\n${versions.join('\n')}`)
-      this.exit()
+      console.log(`All JBrowse versions:\n${versions.join('\n')}`)
+      process.exit(0)
     }
 
-    // mkdir will do nothing if dir exists
-    await fsPromises.mkdir(argsPath, { recursive: true })
+    const localPath = positionals[0]
+    if (!localPath) {
+      console.error('Missing 1 required arg:')
+      console.error('localPath  Location where JBrowse 2 will be installed')
+      console.error('See more help with --help')
+      process.exit(1)
+    }
 
-    if (!force) {
-      await this.checkPath(argsPath)
+    this.debug(`Want to install path at: ${localPath}`)
+
+    // mkdir will do nothing if dir exists
+    await fsPromises.mkdir(localPath, { recursive: true })
+
+    if (!flags.force) {
+      await this.checkPath(localPath)
     }
 
     const locationUrl =
-      url ||
-      (nightly ? await this.getBranch('main') : '') ||
-      (branch ? await this.getBranch(branch) : '') ||
-      (tag ? await this.getTag(tag) : await this.getLatest())
+      flags.url ||
+      (flags.nightly ? await this.getBranch('main') : '') ||
+      (flags.branch ? await this.getBranch(flags.branch) : '') ||
+      (flags.tag ? await this.getTag(flags.tag) : await this.getLatest())
 
-    this.log(`Fetching ${locationUrl}...`)
+    console.log(`Fetching ${locationUrl}...`)
     const response = await fetch(locationUrl)
     if (!response.ok) {
-      this.error(`Failed to fetch: ${response.statusText}`, { exit: 100 })
+      console.error(`Failed to fetch: ${response.statusText}`)
+      process.exit(100)
     }
 
     const type = response.headers.get('content-type')
     if (
-      url &&
+      flags.url &&
       type !== 'application/zip' &&
       type !== 'application/octet-stream'
     ) {
-      this.error(
+      console.error(
         'The URL provided does not seem to be a JBrowse installation URL',
       )
+      process.exit(1)
     }
-    await decompress(Buffer.from(await response.arrayBuffer()), argsPath)
+    await decompress(Buffer.from(await response.arrayBuffer()), localPath)
 
-    this.log(`Unpacked ${locationUrl} at ${argsPath}`)
+    console.log(`Unpacked ${locationUrl} at ${localPath}`)
   }
 
   async checkPath(userPath: string) {
     const allFiles = await fsPromises.readdir(userPath)
     if (allFiles.length > 0) {
-      this.error(
+      console.error(
         `This directory (${userPath}) has existing files and could cause conflicts with create. Please choose another directory or use the force flag to overwrite existing files`,
-        { exit: 120 },
       )
+      process.exit(120)
     }
   }
 
-  async catch(error: unknown) {
-    // @ts-expect-error
-    if (error.parse?.output.flags.listVersions) {
-      const versions = (await this.fetchGithubVersions()).map(
-        version => version.tag_name,
-      )
-      this.log(`All JBrowse versions:\n${versions.join('\n')}`)
-      this.exit()
-    }
-    throw error
+  showHelp() {
+    console.log(`
+${CreateNative.description}
+
+USAGE
+  $ jbrowse create <localPath> [options]
+
+ARGUMENTS
+  localPath  Location where JBrowse 2 will be installed
+
+OPTIONS
+  -h, --help         Show help
+  -f, --force        Overwrites existing JBrowse 2 installation if present in path
+  -l, --listVersions Lists out all versions of JBrowse 2
+  --branch <branch>  Download a development build from a named git branch
+  --nightly          Download the latest development build from the main branch
+  -u, --url <url>    A direct URL to a JBrowse 2 release
+  -t, --tag <tag>    Version of JBrowse 2 to install. Format is v1.0.0. Defaults to latest
+
+EXAMPLES
+${CreateNative.examples.join('\n')}
+`)
   }
 }
